Refatorado TableRow
-- Refatorado TableRow e colocado os componentes elementos da actionCell em seus próprios componentes
-- A TableRow e seus componentes foram colocados em uma pasta _TableRow por questões de hierarquia
